debian/rules: remove override_dh_autoreconf
The Makefile already runs autoreconf -fi in the "utils" target:

  utils:
  	(cd host && \
  	 autoreconf -fi && \
  	 ./configure --prefix=/usr --disable-werror && \
  	 make)

The override is not useful, as it leads to debhelper running autoreconf
as well, but without running ./configure and make in the host directory
afterwards. So autoreconf just runs twice.

I've also considered to change debian/rules to only run the "fw"
target, and not the "utils" target of the Makefile. But that only makes
it more complex, as debhelper would then need to run make twice, once in
the root directory (as "make fw"), and once in the host directory. We
would need to add several lines to debian/rules to do effectively the
same thing.

Make this change now, as "cd host && dh_autoreconf" doesn't work with
debhelper compat level v10 anymore (--sourcedir could be used instead,
but as mentioned above, it's not useful).

Related: OS#5958
Change-Id: I12f379b4ec2de6adc86557d89319ab1d04ed5e73