contrib/jenkins: tweak shell logic
Move the logic down to where make gets called, so we don't need the
variable. Print whether we use CLANG or GCC.

Don't put /opt/llvm-arm/bin infront of PATH unless building with CLANG.
Right now it doesn't seem to have e.g. an override for gcc, but the
files in that path may change when we update
LLVM-embedded-toolchain-for-Arm.

Related: OS#6026
Change-Id: Id768e0dbed9265f042b942e6699683723ca40a7c